feat(vrd): :sparkles: Separate the Random struct and its implementation into a dedicated module file
